refactor(web): consolidate viavai utils into lib utils
