--- web/src/components/viavai/utils.ts
@@ -1,3 +1,0 @@
-export function isObject(value: unknown): value is Record<string, unknown> {
-    return typeof value === 'object' && value !== null;
-}
